requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
Pillow==11.3.0
python-dotenv==1.1.1